# web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "hearsay",
    "description": "Local-first meeting-note transcriber. The core serves this API on `127.0.0.1` only, for the bundled web UI.\n\n**Conventions**\n\n- Base path is `/api`. Every request needs the session token; see the `session_token` scheme.\n- List endpoints return `{ total, page, page_size, items }` and take `page` (at least 1, default 1) and `page_size` (1 to 200, default 50).\n- Errors return `{ \"detail\": \"...\" }`. Database and internal failures render as a literal \"internal error\" — details go to the log only.\n- Times are ISO 8601. Segment `start_s` and `end_s` are seconds relative to the meeting start, not wall clock.\n\n**Not in this document**\n\nThree served routes have no OpenAPI representation and are specified in `docs/api.md`: the live transcript WebSocket at `/ws/meetings/{id}`, meeting audio at `/api/meetings/{id}/audio` (a byte stream with `Range` support), and `/`, which serves the UI with the token injected.",
    "license": {
      "name": "Apache-2.0 WITH Commons-Clause",
      "url": "https://github.com/rwiseman87/hearsay/blob/main/LICENSE"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/meetings": {
      "get": {
        "tags": [
          "meetings"
        ],
        "summary": "List meetings, newest first.",
        "description": "Backs the Library and Dashboard views.",
        "operationId": "list_meetings",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A page of meetings, newest first",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_MeetingRead"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "meetings"
        ],
        "summary": "Start a meeting.",
        "description": "Spawns the capture helper, begins capture, and starts the transcription pipeline. One meeting\nmay be active at a time. `title` is optional; omitted, it defaults to a timestamp-derived name.",
        "operationId": "start_meeting",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MeetingCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Recording started",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeetingRead"
                }
              }
            }
          },
          "409": {
            "description": "A meeting is already recording"
          },
          "422": {
            "description": "Title exceeds 255 characters"
          },
          "503": {
            "description": "Capture engine unavailable"
          }
        }
      }
    },
    "/api/meetings/{id}": {
      "get": {
        "tags": [
          "meetings"
        ],
        "summary": "Read one meeting.",
        "operationId": "get_meeting",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The meeting",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeetingRead"
                }
              }
            }
          },
          "404": {
            "description": "No such meeting"
          }
        }
      },
      "delete": {
        "tags": [
          "meetings"
        ],
        "summary": "Delete a meeting and its recordings.",
        "description": "Removes the database rows (segments, clusters, and both notes tables cascade) and deletes the\non-disk folder. Irreversible. The active recording session cannot be deleted — stop it first.",
        "operationId": "delete_meeting",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Deleted"
          },
          "404": {
            "description": "No such meeting"
          },
          "409": {
            "description": "The meeting is currently recording"
          }
        }
      },
      "patch": {
        "tags": [
          "meetings"
        ],
        "summary": "Rename a meeting.",
        "description": "Updates the title only; filing a meeting into a folder is a separate operation\n(`PUT /api/meetings/{id}/folder`). The title is trimmed before it is stored.",
        "operationId": "update_meeting",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MeetingUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "The renamed meeting",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeetingRead"
                }
              }
            }
          },
          "404": {
            "description": "No such meeting"
          },
          "422": {
            "description": "Blank title, or over 255 characters"
          }
        }
      }
    },
    "/api/meetings/{id}/folder": {
      "put": {
        "tags": [
          "meetings"
        ],
        "summary": "File a meeting into a folder.",
        "description": "A `null` `folder_id` un-files the meeting back to the root. Organizational only — the meeting's\non-disk directory never moves.",
        "operationId": "assign_meeting_folder",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MeetingFolderAssign"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "The re-filed meeting",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeetingRead"
                }
              }
            }
          },
          "404": {
            "description": "The meeting or the target folder is unknown"
          },
          "422": {
            "description": "Malformed request body"
          }
        }
      }
    },
    "/api/meetings/{id}/keep-recording": {
      "post": {
        "tags": [
          "meetings"
        ],
        "summary": "\"Keep recording\": reset the active meeting's inactivity clock so a present-but-quiet user is not\nnudged again or auto-ended. A no-op for any meeting that is not the current recording session\n(404) — there is no clock to reset. Idempotent and cheap; the client calls it when the user\ndismisses the \"still recording?\" prompt.",
        "operationId": "keep_recording",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Silence clock reset"
          },
          "404": {
            "description": "Not the current recording session"
          }
        }
      }
    },
...
    "/api/meetings/{id}/pause": {
      "post": {
        "tags": [
          "meetings"
        ],
        "summary": "Pause the live meeting's capture (the \"Pause\" control): recording + transcription stop and the\ntimeline freezes with no gap until resumed. A 404 for any meeting that is not the current\nrecording session. Idempotent (pausing an already-paused meeting is a no-op 204).",
        "operationId": "pause_meeting",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Paused (or already paused)"
          },
          "404": {
            "description": "Not the current recording session"
          }
        }
      }
    },
...
    "/api/meetings/{id}/resume": {
      "post": {
        "tags": [
          "meetings"
        ],
        "summary": "Resume a paused meeting's capture. A 404 for any meeting that is not the current recording\nsession. Idempotent (resuming a non-paused meeting is a no-op 204).",
        "operationId": "resume_meeting",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Resumed (or already running)"
          },
          "404": {
            "description": "Not the current recording session"
          }
        }
      }
    },
    "/api/meetings/{id}/reveal": {
      "post": {
        "tags": [
          "meetings"
        ],
        "summary": "Open this meeting's recordings folder (audio, transcript, notes, meeting.json) in the OS file\nmanager — the per-meeting analog of the Settings \"Reveal data folder\" action, reusing the same\n`open`-in-Finder helper. Runs in the core (a native process in the user's login session), reached\nover the same-origin HTTP API. Failures surface the reason (not a generic 500) so a broken reveal\nis diagnosable.",
        "operationId": "reveal_meeting",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Handed off to the file manager"
          },
          "404": {
            "description": "No such meeting"
          },
          "503": {
            "description": "The folder could not be opened"
          }
        }
      }
    },
    "/api/meetings/{id}/segments": {
      "get": {
        "tags": [
          "meetings"
        ],
        "summary": "List a meeting's finalized transcript segments, ordered by `start_s`.",
        "description": "How a past meeting reloads from the database, and how a live client recovers after a `resync`\nevent on the transcript WebSocket. Only finalized segments are stored; streamed partials are\nnever persisted.",
        "operationId": "list_segments",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A page of segments, ordered by start time",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_SegmentRead"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/meetings/{id}/stop": {
      "post": {
        "tags": [
          "meetings"
        ],
        "summary": "Stop and finalize a meeting.",
        "description": "Stops capture, flushes the pipeline, rewrites `transcript.md` in timestamp order, and stamps\n`ended_at`. Returns the updated meeting immediately: when a refine or notes step will run the\nstatus is `refining`, and it flips to `finalized` once that background work completes.",
        "operationId": "stop_meeting",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The finalized (or refining) meeting",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeetingRead"
                }
              }
            }
          },
          "404": {
            "description": "No such meeting"
          },
          "503": {
            "description": "Capture engine unavailable"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "MeetingCreate": {
        "type": "object",
        "description": "Start a meeting. `title` defaults to a timestamp-derived name when omitted.",
        "properties": {
          "title": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
    },
    "securitySchemes": {
      "session_token": {
        "type": "http",
        "scheme": "bearer",
        "description": "Per-process session token. `make rust-serve` prints it in the `open:` URL; the packaged app passes it to the webview through a handshake file. Channels that cannot set a header — the WebSocket and the audio element — take `?token=` instead."
      }
    }
  },
  "security": [
    {
      "session_token": []
    }
  ],
...
}